--- backend/src/app.js
@@ -1,12 +1,0 @@
-const express = require('express');
-const cors = require('cors');
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-
-app.get('/health', (req, res) => {
-    res.json({ status: 'ok'});
-});
-
-module.exports = app;
